Fix Visual Studio 2015 compilation issues.
Reviewed-on: https://swiftshader-review.googlesource.com/5780
Tested-by: Nicolas Capens <capn@google.com>

+++ src/Renderer/RoutineCache.hpp
@@ -56,7 +56,7 @@ namespace sw
 
 				precacheDLL = LoadLibrary(dllName);
 				FILE *dir = fopen(dirName, "rb");
-				int ordinal = 1;
+				uintptr_t ordinal = 1;
 
 				while(precacheDLL && dir)
 				{
